Fix minion to not abort chef when salt master not found
Switch from raising an exception to logging a message. Add a conditional wrapper to ensure minion is only configured when salt-master can be found.

This is necessary when configuring a salt master for the first time from a preconfigured machine. The base roles already have the salt-minion code in place, which will abort when the salt master cannot be found. This prevents the chef run from continuing to run and enabling the salt master. It solves a chicken and egg problem.